chore: fix extra lint errors
